fix: fixed some errors

+++ kern/include/syscall.h
@@ -84,7 +84,7 @@ int sys_fork(struct trapframe *ctf, pid_t *retval);
 int sys_execv(const char *program, char **args);
 int sys_read(int fd, userptr_t buf, size_t size, int32_t *retval);
 int sys_write(int fd, userptr_t buf, size_t size, int32_t *retval);
-int sys_lseek(int fd, off_t offset, int whence, int32_t *retval);
+int sys_lseek(int fd, off_t offset, int whence, int64_t *retval);
 int sys_open(userptr_t pathname, int flags, mode_t mode, int32_t *retval);
 int sys_close(int fd);
 int sys_dup2(int oldfd, int newfd, int32_t *retval);
